Add dry run unit test
